search movie_2

+++ threemovieapi/src/main/kotlin/com/threemovie/threemovieapi/domain/movie/service/MovieReviewService.kt
@@ -109,8 +109,8 @@ class MovieReviewService(
 			val movieId = movie_tmp.key
 			for (oneReview in LT_Moviedata) {
 				val member_MovieReview = MovieReview(
-					JSONObject(oneReview.toString()).get("RecommandCount").toString(),
-					JSONObject(oneReview.toString()).get("RegistDate").toString().replace(".", "").trim(),
+					JSONObject(oneReview.toString()).get("RecommandCount").toString().toInt(),
+					JSONObject(oneReview.toString()).get("RegistDate").toString().replace(".", "").trim().toLong(),
 					JSONObject(oneReview.toString()).get("ReviewText").toString(),
 					"LC",
 				)
@@ -206,9 +206,9 @@ class MovieReviewService(
 			
 			for (oneReview in MB_Moviedata) {
 				val member_MovieReview = MovieReview(
-					JSONObject(oneReview.toString()).get("rcmmCnt").toString(),
+					JSONObject(oneReview.toString()).get("rcmmCnt").toString().toInt(),
 					JSONObject(oneReview.toString()).get("fullLstUptDt").toString().split(" ")[0].replace(".", "")
-						.trim(),
+						.trim().toLong(),
 					JSONObject(oneReview.toString()).get("onelnEvalCn").toString(),
 					"MB",
 				)
